Extract ViewModel mocks to test/mock for reuse
